Music.Score.Chord

Description

Provides a representation for chords.

Chord representation

class HasChord a where

Associated Types

type ChordNote a :: *

Methods

getChord :: a -> [ChordNote a]

newtype ChordT a

Constructors

ChordT

Fields

getChordT :: [a]

Voice separation

separateVoices :: Ord a => Score a -> [Score a]

mergePossible :: [Score a] -> [Score a]

Heuristically merge voices if possible

Chord transformations

takeNoteInChord :: Int -> Score b -> Score b

dropNoteInChord :: Int -> Score b -> Score b

takeNotesInChord :: Int -> Score b -> Score b

dropNotesInChord :: Int -> Score b -> Score b

mapSimultaneous :: (Score [a] -> Score [b]) -> Score a -> Score b

Process all simultaneous events.

Two events a and b are considered simultaneous if and only if they have the same era, that is if era a == era b

simultaneous :: Semigroup a => Score a -> Score a

Merge all simultaneous events using their Semigroup instance.

Two events a and b are considered simultaneous if and only if they have the same era, that is if era a == era b

simultaneous' :: Score a -> Score [a]

Group simultaneous events as lists.

Two events a and b are considered simultaneous if and only if they have the same era, that is if era a == era b

Note that simultaneous is identical to simultaneous . fmap return