Music.Score.Export.Lilypond

Description

Lilypond backend

class HasLilypondInstrument a where

Extract instrument info as per Music.Part This is really crude, needs rethinking!

Methods

getLilypondClef :: a -> Int

data Lilypond

A token to represent the Lilypond backend.

data LyContext a

Context passed to the note export. Includes duration and note/rest distinction.

Constructors

LyContext Duration (Maybe a)

type HasLilypond a = (HasBackendNote Lilypond (BackendScoreEvent Lilypond a), HasBackendScore Lilypond a)

Constraint for types that has a Lilypond representation.

Converting to Lilypond

toLilypond :: HasLilypond a => a -> LyMusic

Convert a score to a Lilypond representation.

toLilypondString :: HasLilypond a => a -> String

Convert a score to a Lilypond string.

Lilypond I/O

showLilypond :: HasLilypond a => a -> IO ()

Convert a score to a Lilypond representaiton and print it on the standard output.

openLilypond :: HasLilypond a => a -> IO ()

Typeset a score using Lilypond and open it. (This is simple wrapper around writeLilypond that may not work well on all platforms.)

writeLilypond :: HasLilypond a => FilePath -> a -> IO ()

Convert a score to a Lilypond representation and write to a file.

Customize Lilypond backend

data LilypondOptions

Constructors

LyInlineFormat

LyScoreFormat

openLilypond' :: HasLilypond a => LilypondOptions -> a -> IO ()

Typeset a score using Lilypond and open it. (This is simple wrapper around writeLilypond that may not work well on all platforms.)

writeLilypond' :: HasLilypond a => LilypondOptions -> FilePath -> a -> IO ()

Convert a score to a Lilypond representation and write to a file.