Music.Score.Export.Lilypond2

Documentation

type HasOrdPart a = (HasPart' a, Ord (Part a))

class (HasDynamic' a, HasDynamic a a', HasDynamic a' a'', HasDynamic a a'') => HasDynamic3 a a' a''

type HasDynamicNotation a b c = (HasDynamic3 a b c, Dynamic b ~ Ctxt (Dynamic a), Dynamic c ~ DynamicNotation, Real (Dynamic a))

class Functor (BackendScore b) => HasBackend b where

This class defines types and functions for exporting music. It provides the primitive types and methods used to implement export.

The backend type b is just a type level tag to identify a specific backend. It is typically defined as an empty data declaration.

The actual conversion is handled by the subclasses HasBackendScore and HasBackendNote, which converts the time structure, and the contained music respectively. Thus structure and content are handled separately.

It is often necessary to alter the events based on their surrounding context: for examples the beginning and end of spanners and beams depend on surrounding notes. The BackendContext type allow HasBackendScore instances to provide context for HasBackendNote instances.

 data NoteList

instance HasBackend NoteList where
   type BackendScore NoteList     = []
   type BackendContext NoteList   = Identity
   type BackendNote NoteList      = [(Sum Int, Int)]
   type BackendMusic NoteList     = [(Sum Int, Int)]
   finalizeExport _ = concat

instance HasBackendScore NoteList [a] a where
   exportScore _ = fmap Identity

instance HasBackendNote NoteList a => HasBackendNote NoteList [a] where
   exportNote b ps = mconcat $ map (exportNote b) $ sequenceA ps

instance HasBackendNote NoteList Int where
   exportNote _ (Identity p) = [(mempty ,p)]

instance HasBackendNote NoteList a => HasBackendNote NoteList (DynamicT (Sum Int) a) where
   exportNote b (Identity (DynamicT (d,ps))) = set (mapped._1) d $ exportNote b (Identity ps)
 --

Associated Types

type BackendMusic b :: *

External music representation

type BackendNote b :: *

Notes, chords and rests, with output handled by HasBackendNote

type BackendScore b :: * -> *

Score, voice and time structure, with output handled by HasBackendScore

type BackendContext b :: * -> *

This type may be used to pass context from exportScore to exportNote. If the note export is not context-sensitive, Identity can be used.

Methods

finalizeExport :: b -> BackendScore b (BackendNote b) -> BackendMusic b

class HasBackend b => HasBackendScore b s where

Associated Types

type BackendScoreEvent b s :: *

Methods

exportScore :: b -> s -> BackendScore b (BackendContext b (BackendScoreEvent b s))

class HasBackend b => HasBackendNote b a where

Methods

exportNote :: b -> BackendContext b a -> BackendNote b

exportChord :: b -> BackendContext b [a] -> BackendNote b

export :: (HasBackendScore b s, HasBackendNote b (BackendScoreEvent b s)) => b -> s -> BackendMusic b

Note lists

data NoteList

A simple backend that supports rendering scores to lists of pitch and velocity.

This exists as a sanity check for the Backend classes, and as an example.

toNoteList :: (HasBackendNote NoteList (BackendScoreEvent NoteList s), HasBackendScore NoteList s) => s -> [(Int, Int)]

SuperCollider patterns

data SuperCollider

A token to represent the SuperCollider backend.

type HasSuperCollider a = (HasBackendNote SuperCollider (BackendScoreEvent SuperCollider a), HasBackendScore SuperCollider a)

toSuperCollider :: HasSuperCollider a => a -> String

writeSuperCollider :: HasSuperCollider a => FilePath -> a -> IO ()

openSuperCollider :: HasSuperCollider a => a -> IO ()

MIDI

class HasMidiProgram a

Class of part types with an associated MIDI program number.

data Midi

A token to represent the Midi backend.

type HasMidiNEW a = (HasBackendNote Midi (BackendScoreEvent Midi a), HasBackendScore Midi a)

toMidi :: HasMidiNEW a => a -> Midi

writeMidi :: HasMidiNEW a => FilePath -> a -> IO ()

showMidi :: HasMidiNEW a => a -> IO ()

openMidi :: HasMidiNEW a => a -> IO ()

Lilypond

data Lilypond

A token to represent the Lilypond backend.

type HasLilypondNEW a = (HasBackendNote Lilypond (BackendScoreEvent Lilypond a), HasBackendScore Lilypond a)

Converting to Lilypond

toLilypond :: HasLilypondNEW a => a -> Music

Convert a score to a Lilypond representation.

toLilypondString :: HasLilypondNEW a => a -> String

Convert a score to a Lilypond string.

Lilypond I/O

showLilypond :: HasLilypondNEW a => a -> IO ()

Convert a score to a Lilypond representaiton and print it on the standard output.

openLilypond :: HasLilypondNEW a => a -> IO ()

Typeset a score using Lilypond and open it.

(This is simple wrapper around writeLilypond that may not work well on all platforms.)

writeLilypond :: HasLilypondNEW a => FilePath -> a -> IO ()

Convert a score to a Lilypond representation and write to a file.

Customize Lilypond backend

data LilypondOptions

Constructors

LyInlineFormat

LyScoreFormat

openLilypond' :: HasLilypondNEW a => LilypondOptions -> a -> IO ()

Typeset a score using Lilypond and open it.

(This is simple wrapper around writeLilypond that may not work well on all platforms.)

writeLilypond' :: HasLilypondNEW a => LilypondOptions -> FilePath -> a -> IO ()

Convert a score to a Lilypond representation and write to a file.

MusicXml

data MusicXml

A token to represent the MusicXml backend.

type HasMusicXmlNEW a = (HasBackendNote MusicXml (BackendScoreEvent MusicXml a), HasBackendScore MusicXml a)

Converting to MusicXml

toMusicXml :: HasMusicXmlNEW a => a -> Score

Convert a score to a MusicXml representation.

toMusicXmlString :: HasMusicXmlNEW a => a -> String

Convert a score to a MusicXml string.

MusicXml I/O

showMusicXml :: HasMusicXmlNEW a => a -> IO ()

Convert a score to a MusicXml representaiton and print it on the standard output.

openMusicXml :: HasMusicXmlNEW a => a -> IO ()

Typeset a score using MusicXml and open it.

(This is simple wrapper around writeMusicXml that may not work well on all platforms.)

writeMusicXml :: HasMusicXmlNEW a => FilePath -> a -> IO ()

Convert a score to a MusicXml representation and write to a file.