Music.Score.Export.MusicXml

Description

MusicXml backend

class HasMusicXmlInstrument a where

Extract instrument info as per Music.Part This is really crude, needs rethinking!

Methods

getMusicXmlClef :: a -> Int

data MusicXml

A token to represent the MusicXml backend.

data XmlContext a

Context passed to the note export. Includes duration and note/rest distinction.

Constructors

XmlContext Duration (Maybe a)

type HasMusicXml a = (HasBackendNote MusicXml (BackendScoreEvent MusicXml a), HasBackendScore MusicXml a)

Constraint for types that has a MusicXML representation.

Converting to MusicXml

toMusicXml :: HasMusicXml a => a -> Score

Convert a score to a MusicXML score.

toMusicXmlString :: HasMusicXml a => a -> String

Convert a score to a MusicXML string.

MusicXml I/O

showMusicXml :: HasMusicXml a => a -> IO ()

Convert a score to MusicXML string and print it on the standard output.

openMusicXml :: HasMusicXml a => a -> IO ()

Typeset a score using MusicXML and open it.

(This is simple wrapper around writeMusicXml that may not work well on all platforms.)

writeMusicXml :: HasMusicXml a => FilePath -> a -> IO ()

Convert a score to MusicXML and write to a file.