Music.Score.Export.NoteList

Description

A simple backend that supports rendering scores to lists of pitch and velocity.

This exists as a sanity check for the Backend classes, and as an example.

Note list backend

data NoteList

A token to represent the note list backend.

toNoteList :: (HasBackendNote NoteList (BackendScoreEvent NoteList s), HasBackendScore NoteList s) => s -> [(Int, Int)]

Export music as a note list.