Music.Score.Harmonics

Description

Provides a representation of harmonics.

Harmonics

class HasHarmonic a where

Methods

setNatural :: Bool -> a -> a

setHarmonic :: Int -> a -> a

newtype HarmonicT a

Constructors

HarmonicT

Fields

getHarmonicT :: Couple (Any, Sum Int) a

harmonic :: HasHarmonic a => Int -> a -> a

Make all notes natural harmonics on the given overtone (1 for octave, 2 for fifth etc). Sounding pitch is unaffected, but notated output is transposed automatically.

artificial :: HasHarmonic a => a -> a

Make all notes natural harmonics on the given overtone (1 for octave, 2 for fifth etc). Sounding pitch is unaffected, but notated output is transposed automatically.