Music.Score.Import.Midi

Description

Provides MIDI import.

Warning Experimental module.

Documentation

type IsMidi a = (IsPitch a, HasPart' a, Ord (Part a), Enum (Part a), Num (Pitch a), HasTremolo a, HasArticulation a a, Tiable a)

This constraint includes all note types that can be constructed from a Midi representation.

fromMidi :: IsMidi a => Midi -> Score a

Convert a score from a Midi representation.

readMidi :: IsMidi a => FilePath -> IO (Score a)

Read a Midi score from a file. Fails if the file could not be read or if a parsing error occurs.

readMidiMaybe :: IsMidi a => FilePath -> IO (Maybe (Score a))

Read a Midi score from a file. Fails if the file could not be read, and returns Nothing if a parsing error occurs.

readMidiEither :: IsMidi a => FilePath -> IO (Either String (Score a))

Read a Midi score from a file. Fails if the file could not be read, and returns Left m if a parsing error occurs.