Music.Score.Internal.Util

Documentation

divideList :: Int -> [a] -> [[a]]

Divide a list into parts of maximum length n. > category : List > depends : base

splitWhile :: (a -> Bool) -> [a] -> [[a]]

Group a list into sublists whereever a predicate holds. The matched element is the first in the sublist.

splitWhile isSpace "foo bar baz"
   ===> ["foo"," bar"," baz"]

splitWhile (> 3) [1,5,4,7,0,1,2]
   ===> [[1],[5],[4],[7,0,1,2]]

category : List
depends : base

breakList :: Int -> [a] -> [a] -> [a]

Break up a list into parts of maximum length n, inserting the given list as separator. Useful for breaking up strings, as in breakList 80 "n" str.

category : List
depends : base

mapIndexed :: (Int -> a -> b) -> [a] -> [b]

Map over the indices and elements of list. > category : List > depends : base

dup :: a -> (a, a)

Duplicate an element. > category: Combinator, Tuple > depends: base

unf :: (a -> Maybe a) -> a -> [a]

Unfold a partial function. This is a simpler version of unfoldr. > category: Function, List > depends: base

mapF :: (b -> b) -> [b] -> [b]

Map over first elements of a list. Biased on first element for shorter lists. > category: List > depends: base

mapT :: (b -> b) -> [b] -> [b]

Map over all but the first and last elements of a list. Biased on middle elements for shorter lists. > category: List > depends: base

mapL :: (b -> b) -> [b] -> [b]

Map over last elements of a list. Biased on last element for shorter lists. > category: List > depends: base

mapFTL :: (a -> b) -> (a -> b) -> (a -> b) -> [a] -> [b]

Map over first, middle and last elements of list. Biased on first, then on first and last for short lists.

category: List
depends: base

filterOnce :: (a -> Bool) -> [a] -> [a]

Extract the first consecutive sublist for which the predicate returns true, or the empty list if no such sublist exists. > category: List > depends: base

rots :: [a] -> [[a]]

Returns all rotations of the given list. Given an infinite list, returns an infinite list of rotated infinite lists. > category: List > depends: base

rotl :: [a] -> [a]

category: List
depends: base

rotr :: [a] -> [a]

category: List
depends: base

rotated :: Int -> [a] -> [a]

category: List
depends: base

curry3 :: ((a, b, c) -> d) -> a -> b -> c -> d

uncurry3 :: (a -> b -> c -> d) -> (a, b, c) -> d

untripl :: (a, b, c) -> ((a, b), c)

tripl :: ((a, b), c) -> (a, b, c)

tripr :: (a, (b, c)) -> (a, b, c)

partial2 :: (a -> b -> Bool) -> a -> b -> Maybe b

partial3 :: (a -> b -> c -> Bool) -> a -> b -> c -> Maybe c

list :: r -> ([a] -> r) -> [a] -> r

Case matching on lists. > category: List > depends: base

merge :: Ord a => [a] -> [a] -> [a]

Merge lists. > category: List > depends: base

mergeBy :: (a -> a -> Ordering) -> [a] -> [a] -> [a]

Merge lists. > category: List > depends: base

mergeBy' :: (a -> a -> Bool) -> [a] -> [a] -> [a]

composed :: [b -> b] -> b -> b

Compose all functions. > category: Function > depends: base

unRatio :: Integral a => Ratio a -> (a, a)

Separate a ratio. > category: Math > depends: base

showRatio :: (Integral a, Show a) => Ratio a -> String

Nicer printing of ratio as ordinary fractions. > category: Math > depends: base

retainUpdates :: Eq a => [a] -> [Maybe a]

replic :: Integral a => a -> b -> [b]

swap :: (a, b) -> (b, a)

withNext :: [a] -> [(a, Maybe a)]

withPrev :: [a] -> [(Maybe a, a)]

withPrevNext :: [a] -> [(Maybe a, a, Maybe a)]

mapWithNext :: (a -> Maybe a -> b) -> [a] -> [b]

mapWithPrev :: (Maybe a -> a -> b) -> [a] -> [b]

mapWithPrevNext :: (Maybe a -> a -> Maybe a -> b) -> [a] -> [b]

toDouble :: Real a => a -> Double

through :: Applicative f => Lens' s a -> Lens s t a b -> Lens (f s) (f t) (f a) (f b)

single :: Prism' [a] a

floor' :: RealFrac a => a -> a

inspecting :: Eq a => (b -> a) -> b -> b -> Bool