Music.Score.Meta.Key

Description

Provides key signature meta-data.

Key signature type

data Fifths

data KeySignature

A key signature, represented by number of fifths from C and mode.

key :: Fifths -> Bool -> KeySignature

Create a major or minor signature.

isMajorKey :: KeySignature -> Bool

isMinorKey :: KeySignature -> Bool

Adding key signatures to scores

keySignature :: (HasMeta a, HasPosition a) => KeySignature -> a -> a

Set the key signature of the given score.

keySignatureDuring :: HasMeta a => Span -> KeySignature -> a -> a

Set the key signature of the given part of a score.

Extracting key signatures

withKeySignature :: KeySignature -> (KeySignature -> Score a -> Score a) -> Score a -> Score a

Extract all key signatures from the given score, using the given default key signature.