Music.Score.Meta.Time

Description

Provides time signatures and related meta-data.

Time signature type

data TimeSignature

A time signature is a sequence of beat numbers and a note value (i.e. an expression on the form (a1+a2...)/b). For simple time signatures just one beat number is used.

TimeSignature is an instance of Fractional and can be used as follows:

timeSignature (4/4)
timeSignature (6/8)
timeSignature ((3+2)/4)

time :: Integer -> Integer -> TimeSignature

Create a simple time signature.

compoundTime :: [Integer] -> Integer -> TimeSignature

Create a compound time signature.

isSimpleTime :: TimeSignature -> Bool

Whether this is a simple time signature.

isCompoundTime :: TimeSignature -> Bool

Whether this is a compound time signature.

toSimpleTime :: TimeSignature -> TimeSignature

Convert to a simple time signature by adding all numerators. If given a simple time signature, returns it.

getTimeSignature :: TimeSignature -> ([Integer], Integer)

Extract the components of a time signature. Semantic function.

Typically used with the ViewPatterns extension, as in

foo (getTimeSignature -> (beats, noteValue)) = ...

Adding time signature to scores

timeSignature :: (HasMeta a, HasPosition a, Transformable a) => TimeSignature -> a -> a

Set the time signature of the given score.

timeSignatureDuring :: HasMeta a => Span -> TimeSignature -> a -> a

Set the time signature of the given part of a score.

Extracting time signatures

withTimeSignature :: TimeSignature -> (TimeSignature -> Score a -> Score a) -> Score a -> Score a

Extract the time signature from the given score, using the given default time signature.

getTimeSignatures :: TimeSignature -> Score a -> Reactive TimeSignature

getTimeSignatureChanges :: TimeSignature -> Score a -> [(Time, TimeSignature)]

Utility

getBarDurations :: [(TimeSignature, Duration)] -> [Duration]

Given a list of time signatures and the duration between them (TODO use voice), return a list of appropriate bar durations.

getBarTimeSignatures :: [(TimeSignature, Duration)] -> [TimeSignature]

Given a list of time signatures and the duration between them (TODO use voice), return a list of appropriate time signatures for each bar.

standardTimeSignature :: Duration -> TimeSignature

Time signature typically used for the given duration.