Music.Score.Ornaments

Description

Provides functions for manipulating ornaments (and miscellaneous stuff to be given its own module soon...).

Tremolo

class HasTremolo a where

Methods

setTrem :: Int -> a -> a

newtype TremoloT a

Constructors

TremoloT

Fields

getTremoloT :: Couple (Max Word) a

tremolo :: HasTremolo a => Int -> a -> a

Set the number of tremolo divisions for all notes in the score.

Text

class HasText a where

Methods

addText :: String -> a -> a

newtype TextT a

Constructors

TextT

Fields

getTextT :: Couple [String] a

text :: (HasPhrases' s a, HasText a) => String -> s -> s

Attach the given text to the first note in the score.

Harmonics

class HasHarmonic a where

Methods

setNatural :: Bool -> a -> a

setHarmonic :: Int -> a -> a

newtype HarmonicT a

Constructors

HarmonicT

Fields

getHarmonicT :: Couple (Any, Sum Int) a

harmonic :: HasHarmonic a => Int -> a -> a

Make all notes natural harmonics on the given overtone (1 for octave, 2 for fifth etc). Sounding pitch is unaffected, but notated output is transposed automatically.

artificial :: HasHarmonic a => a -> a

Make all notes natural harmonics on the given overtone (1 for octave, 2 for fifth etc). Sounding pitch is unaffected, but notated output is transposed automatically.

Slides and glissando

class HasSlide a where

Methods

setBeginGliss :: Bool -> a -> a

setBeginSlide :: Bool -> a -> a

setEndGliss :: Bool -> a -> a

setEndSlide :: Bool -> a -> a

newtype SlideT a

Constructors

SlideT

Fields

getSlideT :: Couple ((Any, Any), (Any, Any)) a

slide :: (HasPhrases' s a, HasSlide a) => s -> s

Add a slide between the first and the last note.

glissando :: (HasPhrases' s a, HasSlide a) => s -> s

Add a glissando between the first and the last note.