Music.Score.Phrases

Description

This module provides phrase-wise traversal.

Warning This module is experimental.

HasPhrases class

class HasPhrases s t a b | s -> a, t -> b, s b -> t, t a -> s where

Classes that provide a phrase traversal.

Methods

mvoices :: Traversal s t (MVoice a) (MVoice b)

type HasPhrases' s a = HasPhrases s s a a

phrases :: HasPhrases s t a b => Traversal s t (Phrase a) (Phrase b)

A generic phrase-traversal.

phrases' :: HasPhrases' s a => Traversal' s (Phrase a)

A simple generic phrase-traversal.

Phrase types etc

type Phrase a = Voice a

For a phrase, we simply use a voice without rests.

To represent a sequence of phrases we provide two equivalent representations:

MVoice is a sequence of notes/chords or rests. All consecutive non-rests consitute a phrase.

PVoice is a sequence of phrases or durations.

type MVoice a = Voice (Maybe a)

A sequence of phrases or rests, represented as notes or rests.

Each consecutive sequence of non-rest elements is considered to be a phrase. For an explicit representation of the phrase structure, see mvoicePVoice.

type PVoice a = [Either Duration (Phrase a)]

A sequence of phrases or rests, represented with explicit phrase structure.

Utility

mvoicePVoice :: Lens (MVoice a) (MVoice b) (PVoice a) (PVoice b)

View an MVoice as a PVoice.

unsafeMvoicePVoice :: Iso (MVoice a) (MVoice b) (PVoice a) (PVoice b)

View an MVoice as a PVoice and vice versa.

This a valid Iso up to meta-data equivalence.

singleMVoice :: Prism (Score a) (Score b) (MVoice a) (MVoice b)