Music.Score.Pitch

Description

Provides functions for manipulating pitch.

Pitch type

type family Pitch s :: *

This type fuction is used to access the pitch type for a given type.

type family SetPitch b s :: *

This type fuction is used to update the pitch type for a given type. The first argument is the new type.

type Interval a = Diff (Pitch a)

Associated interval type.

HasPitch classes

class HasPitches s t => HasPitch s t where

Class of types that provide a single pitch.

Methods

pitch :: Lens s t (Pitch s) (Pitch t)

Access the pitch.

As this is a Traversal, you can use all combinators from the lens package, for example:

  pitch .~ c    :: (HasPitch' a, IsPitch a)      => a -> a
  pitch +~ 2    :: (HasPitch' a, Num (Pitch a))  => a -> a
  pitch %~ succ :: (HasPitch' a, Enum (Pitch a)) => a -> a
  view pitch    :: HasPitches' a                 => a -> Pitch a
  set  pitch    :: HasPitches a b                => Pitch b -> a -> b
  over pitch    :: HasPitches a b                => (Pitch a -> Pitch b) -> a -> b
  

class (Transformable (Pitch s), Transformable (Pitch t), SetPitch (Pitch t) s ~ t) => HasPitches s t where

Class of types that provide zero or more pitches.

Methods

pitches :: Traversal s t (Pitch s) (Pitch t)

Access all pitches.

As this is a Traversal, you can use all combinators from the lens package, for example:

  toListOf pitches  :: HasPitches' a                  => a -> [Pitch a]
  allOf pitches     :: (HasPitches' a)                => (Pitch a -> Bool) -> a -> Bool
  maximumOf pitches :: (HasPitches' a, Ord (Pitch a)) => a -> Maybe (Pitch a)
  set  pitches      :: HasPitches a b                 => Pitch b -> a -> b
  over pitches      :: HasPitches a b                 => (Pitch a -> Pitch b) -> a -> b
  

Simple versions

type HasPitch' a = HasPitch a a

type HasPitches' a = HasPitches a a

pitch' :: HasPitch' s => Lens' s (Pitch s)

Access the pitch.

Same as pitch, but without polymorphic update.

pitches' :: HasPitches' s => Traversal' s (Pitch s)

Access all pitches.

Same as pitches, but without polymorphic update.

Transposition

type PitchPair v w = (Num (Scalar v), IsInterval v, IsPitch w)

type AffinePair v w = (VectorSpace v, AffineSpace w)

type Transposable a = (HasPitches' a, AffinePair (Interval a) (Pitch a), PitchPair (Interval a) (Pitch a))

Class of types that can be transposed, inverted and so on.

up :: Transposable a => Interval a -> a -> a

Transpose (translate) up.

>>> up m3 (c :: Pitch)
eb

>>> up _P5 [c,d,e :: Pitch]
[g,a,b]

>>> up _P5 [440 :: Hertz, 442, 810]
[g,a,b]

down :: Transposable a => Interval a -> a -> a

Transpose (translate) down.

>>> down m3 (c :: Pitch)
a

>>> down _P5 [c,d,e]
[f_,g_,a_]

above :: (Semigroup a, Transposable a) => Interval a -> a -> a

Add the given interval above.

>>> above _P8 [c :: Pitch]
[c,c']

below :: (Semigroup a, Transposable a) => Interval a -> a -> a

Add the given interval below.

>>> below _P8 [c :: Pitch]
[c,c_]

octavesUp :: Transposable a => Scalar (Interval a) -> a -> a

Transpose up by the given number of octaves.

>>> octavesUp 2 (c :: Pitch)
c''

>>> octavesUp 1 [c,d,e]
[c',d',e']

>>> octavesUp (-1) [c,d,e]
[c_,d_,e_]

octavesDown :: Transposable a => Scalar (Interval a) -> a -> a

Transpose down by the given number of octaves.

>>> octavesDown 2 (c :: Pitch)
c__

>>> octavesDown 1 [c,d,e]
[c_,d_,e_]

>>> octavesDown (-1) [c,d,e]
[c',d',e']

_15va :: Transposable a => a -> a

Same as octavesUp 2.

_8va :: Transposable a => a -> a

Same as octavesUp 1.

_8vb :: Transposable a => a -> a

Same as octavesDown 1.

_15vb :: Transposable a => a -> a

Same as octavesDown 2.

Inversion

invertPitches :: Transposable a => Pitch a -> a -> a

Invert pitches.

Folds

highest :: (HasPitches' a, Ord (Pitch a)) => a -> Maybe (Pitch a)

Extract the highest pitch. Returns Nothing if there are none.

lowest :: (HasPitches' a, Ord (Pitch a)) => a -> Maybe (Pitch a)

Extract the lowest pitch. Returns Nothing if there are none.

meanPitch :: (HasPitches' a, Fractional (Pitch a)) => a -> Maybe (Pitch a)

Extract the average pitch. Returns Nothing if there are none.