Music.Time.Behavior

Description

Behavior type

data Behavior a

A Behavior is a value varying over time.

Use focusing to view a particular segment.

Examples

behavior :: Iso (Time -> a) (Time -> b) (Behavior a) (Behavior b)

View a behavior as a time function and vice versa.

Note that this is just an alias defined to make the documentation nicer:

behavior = tabulated

Combinators

switch :: Time -> Behavior a -> Behavior a -> Behavior a

Instantly switch from one behavior to another.

switch' :: Time -> Behavior a -> Behavior a -> Behavior a -> Behavior a

Instantly switch from one behavior to another with an optinal intermediate value.

trimBefore :: Monoid a => Time -> Behavior a -> Behavior a

Replace everthing before the given time by mempty.

trimAfter :: Monoid a => Time -> Behavior a -> Behavior a

Replace everthing after the given time by mempty.

Common behaviors

Oscillators

line :: Fractional a => Behavior a

A behavior that gives the current time.

Should really have the type Behavior Time, but is provided in a more general form for convenience.

sawtooth :: RealFrac a => Behavior a

A behavior that goes from 0 to 1 repeatedly with a period of 1.

sine :: Floating a => Behavior a

A behavior that

cosine :: Floating a => Behavior a

A behavior that

Impulse functions

unit :: Fractional a => Behavior a

A behavior that varies from 0 to 1 during the same time interval and is 0 before and 1 after that interval.

impulse :: Num a => Behavior a

A behavior that is 1 at time 0, and 0 at all other times.

turnOn :: Behavior Integer

A behavior that goes from 0 to 1 at time 0.

turnOff :: Behavior Integer

A behavior that goes from 1 to 0 at time 0.