Music.Time.Bound

Description

Warning This module is experimental.

Bound type

newtype Bound a

Bound restricts the start and stop time of a value, and prevents access to values outside the bounds.

Bound is especially useful to restrict the range of a Behavior. If we have a value with can only be reasonably defined for a particular time range, we can represent it as Bound Behavior. This is isomorphic to a Note Segment, and bounded whitnesses the isomorphism.

Bound is not Foldable or Traversable, as that would allow us to access values outside the bounds. However, we can still access values of a Bound Behavior in a safe manner using trim or splice.

The semantics are given by

 type Bound a = (Time, Time, a)

Constructors

Bound

Fields

getBound :: (Span, a)

Construction

bounds :: Time -> Time -> a -> Bound a

Add bounds.

bounding :: Span -> a -> Bound a

Add bounds.

 (s,x)^.note = (bounding s . transform s) x