Music.Time.Delayable

Description

Provides delayable values.

Delayable class

class Delayable a where

Delayable values.

Methods

delay :: Duration -> a -> a

Delay a value.

undelay :: Delayable a => Duration -> a -> a

delaying :: (Delayable a, Delayable b) => Duration -> (a -> b) -> a -> b

Apply a function under delay. See also sunder.

move :: Delayable a => Duration -> a -> a

Move a score forward in time. Equivalent to delay.

moveBack :: Delayable a => Duration -> a -> a

Move a score backward in time. Negated verison of delay.

Utility

delayTime :: Delayable a => Time -> a -> a

Delay relative to origin. Provided for situations when you have a value that should forward based on the distance between some time t and the origin, but it does not necessarily have a start time.

newtype NoDelay a

Constructors

NoDelay

Fields

getNoDelay :: a