Music.Time.Delayed

Description

Delayed type

data Delayed a

Delayed represents a value with an offset in time.

A delayed value has a known position, but no duration.

Placing a value inside Delayed does not make it invariant under stretch, as the offset of a delayed value may be stretched with respect to the origin. However, in contrast to a note the duration is not stretched.

Construction

delayed :: Iso (Time, a) (Time, b) (Delayed a) (Delayed b)

View a delayed value as a pair of a the original value and a delay time.

Inspecting delayed values

delayedValue :: (Transformable a, Transformable b) => Lens (Delayed a) (Delayed b) a b

View a delayed value as a pair of the original value and the transformation (and vice versa).