Music.Time.Internal.Quantize

Description

Provides rhythmical quantization.

You generally need not worry about this module, as quantization is handled automatically by HasBackendScore instances.

Warning This modules is poorly named and will probably be split or moved soon.

Rhythm type

data Rhythm a

Constructors

Beat Duration a

Group [Rhythm a]

Dotted Int (Rhythm a)

Tuplet Duration (Rhythm a)

mapWithDur :: (Duration -> a -> b) -> Rhythm a -> Rhythm b

Quantization

quantize :: Tiable a => [(Duration, a)] -> Either String (Rhythm a)

rewrite :: Rhythm a -> Rhythm a

dotMod :: Int -> Duration

Return the scaling applied to a note with the given number of dots (i.e. 32, 74 etc).

Utility

drawRhythm :: Show a => Rhythm a -> String