Music.Time.Internal.Transform

Description

Documentation

module Music.Time.Types

The Transformable class

class Transformable a where

Class of values that can be transformed (i.e. scaled and moved) in time.

Methods

transform :: Span -> a -> a

transformed :: (Transformable a, Transformable b) => Span -> Iso a b a b

View the given value in the context of the given transformation.

Apply under a transformation

whilst :: (Transformable a, Transformable b) => (a -> b) -> Span -> a -> b

Apply a function under transformation.

>>> stretch 2 `whilst` delaying 2 $ (1 <-> 2)
4 <-> 6

whilstL :: (Functor f, Transformable a, Transformable b) => LensLike f s t a b -> LensLike f (Span, s) (Span, t) a b

whilstLT :: (Functor f, Transformable a, Transformable b) => LensLike f s t a b -> LensLike f (Time, s) (Time, t) a b

whilstLD :: (Functor f, Transformable a, Transformable b) => LensLike f s t a b -> LensLike f (Duration, s) (Duration, t) a b

onSpan :: (Transformable s, Transformable t, Functor f) => LensLike f s t a b -> Span -> LensLike f s t a b

Transforms a lens of to a Transformable type to act inside a transformation.

Designed to be used infix, as in

l onSpan (2 <-> 3)

Specific transformations

Transformations

delaying :: Duration -> Span

A transformation that moves a value forward in time.

undelaying :: Duration -> Span

A transformation that moves a value backward in time.

stretching :: Duration -> Span

A transformation that stretches (augments) a value by the given factor.

compressing :: Duration -> Span

A transformation that compresses (diminishes) a value by the given factor.

Transforming values

delay :: Transformable a => Duration -> a -> a

Moves a value forward in time.

undelay :: Transformable a => Duration -> a -> a

Moves a value backward in time. Equnitvalent to stretch . negate.

stretch :: Transformable a => Duration -> a -> a

Stretches (augments) a value by the given factor.

compress :: Transformable a => Duration -> a -> a

Compresses (diminishes) a score. Equnitvalent to stretch . recip.

delayTime :: Transformable a => Time -> a -> a

Delay relative to origin.

Provided for situations when we really want to use startAt, but the type does not have an instance for HasPosition and we can assume that the value is starting at time zero.