Music.Time.Past

Documentation

newtype Past a

Past represents a value occuring up to some point in time.

It may be seen as a note whose era is a left-open time interval.

Constructors

Past

Fields

getPast :: (a, Time)

past :: Past a -> Time -> Maybe a

Query a past value.

pastSeg :: Past (Segment a) -> Behavior (Maybe a)

Project a segment (backwards) up to the given point.