Music.Time.Placed

Description

Placed type

data Placed a

Placed represents a value with an offset in time.

A placed value has a known position, but no duration.

Placing a value inside Placed does not make it invariant under stretch, as the offset of a placed value may be stretched with respect to the origin. However, in contrast to a note the duration is not stretched.

Construction

placed :: Iso (Time, a) (Time, b) (Placed a) (Placed b)

placee :: (Transformable a, Transformable b) => Lens (Placed a) (Placed b) a b