Music.Time.Position

Description

This misleadingly named module provide a way to query a value for its duration, onset and offset.

Documentation

module Music.Time.Duration

The HasPosition class

class HasDuration a => HasPosition a where

Class of values that have a position in time.

Many values such as notes, envelopes etc can in fact have many positions such as onset, attack point, offset, decay point time etc. Rather than having separate methods for a discrete set of cases, this class provides an interpolation from a local position to a global position. While the local position goes from 0 to 1, the global position goes from the onset to the offset of the value.

For instantaneous values, a suitable instance is:

_position x = const t

For values with an onset and offset we can use alerp:

_position x = alerp (_onset x) (_offset x)

Minimal complete definition

Nothing

Methods

_position :: a -> Duration -> Time

Return the onset of the given value, or the value between the attack and decay phases.

_onset, _offset :: a -> Time

Return the onset of the given value, or the value between the attack and decay phases.

Era

era :: (HasPosition a, Transformable a) => Lens' a Span

A lens to the position

Position

position :: (HasPosition a, Transformable a) => Duration -> Lens' a Time

Position of the given value.

Specific positions

onset :: (HasPosition a, Transformable a) => Lens' a Time

Onset of the given value.

midpoint :: (HasPosition a, Transformable a) => Lens' a Time

Midpoint of the given value, or the value between the decay and sustain phases.

offset :: (HasPosition a, Transformable a) => Lens' a Time

Onset of the given value.

preOnset :: (HasPosition a, Transformable a) => Lens' a Time

Pre-onset of the given value, or the value right before the attack phase.

postOffset :: (HasPosition a, Transformable a) => Lens' a Time

Post-offset of the given value, or the value right after the release phase.

postOnset :: (HasPosition a, Transformable a) => Lens' a Time

Deprecated: Use midpoint

Moving

startAt :: (Transformable a, HasPosition a) => Time -> a -> a

Move a value forward in time.

stopAt :: (Transformable a, HasPosition a) => Time -> a -> a

Move a value forward in time.

placeAt :: (Transformable a, HasPosition a) => Duration -> Time -> a -> a

Align a value to a given position.

placeAt p t places the given thing so that its position p is at time t

placeAt 0 = startAt
placeAt 1 = stopAt

Internal

_setEra :: (HasPosition a, Transformable a) => Span -> a -> a

Place a value over the given span.

placeAt s t places the given thing so that x^.place = s

_getEra :: HasPosition a => a -> Span