Music.Time.Reactive

Description

Reactive type

data Reactive a

Forms an applicative as per Behavior, but only switches at discrete points.

Query

initial :: Reactive a -> a

Get the initial value.

final :: Reactive a -> a

Get the final value.

intermediate :: Transformable a => Reactive a -> [Event a]

Get all intermediate values.

discrete :: Reactive a -> Behavior a

Realize a Reactive value as a discretely changing behavior.

updates :: Reactive a -> [(Time, a)]

Get the time of all updates and the value switched to at this point.

occs :: Reactive a -> [Time]

atTime :: Reactive a -> Time -> a

Construction

Combine

switchR :: Time -> Reactive a -> Reactive a -> Reactive a

switch t a b behaves as a before time t, then as b.

trimR :: Monoid a => Span -> Reactive a -> Reactive a

Split

splitReactive :: Reactive a -> Either a ((a, Time), [Event a], (Time, a))

Split a reactive into events, as well as the values before and after the first/last update

Sampling

continous :: Reactive (Segment a) -> Behavior a

Realize a Reactive value as an continous behavior.

continousWith :: Segment (a -> b) -> Reactive a -> Behavior b

Realize a Reactive value as an continous behavior.

sample :: [Time] -> Behavior a -> Reactive a

Sample a Behavior into a reactive.