Music.Time.Rest

Description

Many time structures such as Score allows for rests between notes. Generally rests are simply treated as blank space, and thus have no duration. Sometimes it is useful to represent rests explicitly, so this module provides an alias for pure Nothing that can be used to that end.

To remove rests from a score, use mcatMaybes, for example:

 open $ mcatMaybes $ scat [c,d,rest^*2,e]^/8

Rests

rest :: Applicative f => f (Maybe a)