Music.Time.Reverse

Description

Provides reversible values.

Documentation

module Music.Time.Position

The Reversible class

class Transformable a => Reversible a where

Class of values that can be reversed (retrograded).

For positioned values succh as Note, the value is reversed relative to its middle point, i.e. the onset value becomes the offset value and vice versa.

For non-positioned values such as Stretched, the value is reversed in-place.

FIXME Second law is incompatible with revDefault (and the Span definition below)

Law

rev (rev a) = a

abs (x^.duration) = (rev x)^.duration

rev s transform a = rev (s transform a)

or equivalently,

transform . rev = fmap rev . transform

For Span

rev = over range swap

Methods

rev :: a -> a

Reverse (retrograde) the given value.

Reversing

reversed :: Reversible a => Iso' a a

View the reverse of a value.

>>> [1,2,3] & reversed %~ Data.List.sort
[3,2,1]

revDefault :: (HasPosition a, Transformable a) => a -> a

A default implementation of rev

Utility

newtype NoReverse a

Constructors

NoReverse

Fields

getNoReverse :: a