Music.Time.Score

Description

Score type

data Score a

A Score is a sequential or parallel composition of values, and allows overlapping events

You typically create a Score using score, notes, voices, and phrases, or the Alternative interface.

Score is an instance of Transformable, so you can use delay and stretch.

Score is an instance of HasPosition, so you can use duration, onset, offset, era.

To inspect or deconstruct a score, see notes, voices, and phrases, as well as singleNote, singleVoice, and singlePhrase

The semantics are given by

 type Score a = [Note a]

Construction

score :: Getter [Note a] (Score a)

Create a score from a list of notes.

This is a getter (rather than a function) for consistency:

 [ (0 <-> 1, 10)^.note,
   (1 <-> 2, 20)^.note,
   (3 <-> 4, 30)^.note ]^.score

 view score $ map (view note) [(0 <-> 1, 1)]

Se also notes.

Extracting values

notes :: Lens (Score a) (Score b) [Note a] [Note b]

View a Score as a list of Note values.

 view notes                        :: Score a -> [Note a]
 set  notes                        :: [Note a] -> Score a -> Score a
 over notes                        :: ([Note a] -> [Note b]) -> Score a -> Score b

 preview  (notes . each)           :: Score a -> Maybe (Note a)
 preview  (notes . element 1)      :: Score a -> Maybe (Note a)
 preview  (notes . elements odd)   :: Score a -> Maybe (Note a)

 set      (notes . each)           :: Note a -> Score a -> Score a
 set      (notes . element 1)      :: Note a -> Score a -> Score a
 set      (notes . elements odd)   :: Note a -> Score a -> Score a

 over     (notes . each)           :: (Note a -> Note b) -> Score a -> Score b
 over     (notes . element 1)      :: (Note a -> Note a) -> Score a -> Score a
 over     (notes . elements odd)   :: (Note a -> Note a) -> Score a -> Score a

 toListOf (notes . each)                :: Score a -> [Note a]
 toListOf (notes . elements odd)        :: Score a -> [Note a]
 toListOf (notes . each . filtered
              (\x -> _duration x < 2))  :: Score a -> [Note a]

This is not an Iso, as the note list representation does not contain meta-data. To construct a score from a note list, use score or flip (set notes) empty.

events :: Lens (Score a) (Score b) [(Time, Duration, a)] [(Time, Duration, b)]

View a score as a list of events, i.e. time-duration-value triplets.

This is a convenient combination of notes and event.

 events = notes . through event event

Pattern matching

singleNote :: Prism' (Score a) (Note a)

View a score as a single note.

Unsafe versions

unsafeNotes :: Iso (Score a) (Score b) [Note a] [Note b]

View a score as a list of notes.

This only an isomorphism up to meta-data. See also the safe (but more restricted) notes and score.

unsafeEvents :: Iso (Score a) (Score b) [(Time, Duration, a)] [(Time, Duration, b)]

View a score as a list of events.

This only an isomorphism up to meta-data. See also the safe (but more restricted) notes and score.

Simultaneous values

simult :: Transformable a => Lens (Score a) (Score b) (Score [a]) (Score [b])

simultaneous :: (Transformable a, Semigroup a) => Score a -> Score a

Merge all simultaneous events using their Semigroup instance.

Two events a and b are considered simultaneous if and only if they have the same era, that is if era a == era b

Traversing

mapWithSpan :: (Span -> a -> b) -> Score a -> Score b

Map over the values in a score.

filterWithSpan :: (Span -> a -> Bool) -> Score a -> Score a

Filter the values in a score.

mapFilterWithSpan :: (Span -> a -> Maybe b) -> Score a -> Score b

Combination of mapEvents and filterEvents.

mapEvents :: (Time -> Duration -> a -> b) -> Score a -> Score b

Map over the values in a score.

filterEvents :: (Time -> Duration -> a -> Bool) -> Score a -> Score a

Filter the values in a score.

mapFilterEvents :: (Time -> Duration -> a -> Maybe b) -> Score a -> Score b

Efficient combination of mapEvents and filterEvents.