Music.Time.Segment

Description

Behavior type

data Behavior a

A Behavior is a value varying over time.

Use focusing to view a particular Segment.

The semantics are given by

 type Behavior a = Time -> a

Examples

behavior :: Iso (Time -> a) (Time -> b) (Behavior a) (Behavior b)

View a behavior as a time function and vice versa.

Note that this is just an alias defined to make the documentation nicer:

 behavior = tabulated

Combinators

switch :: Time -> Behavior a -> Behavior a -> Behavior a

Instantly switch from one behavior to another.

switch' :: Time -> Behavior a -> Behavior a -> Behavior a -> Behavior a

Instantly switch from one behavior to another with an optinal intermediate value.

splice :: Behavior a -> Bound (Behavior a) -> Behavior a

Inserts a bounded behavior on top of another behavior.

 trim = splice mempty

(Named after the analogous tape-editing technique.)

trim :: Monoid b => Bound (Behavior b) -> Behavior b

Extract a bounded behavior, replacing all values outside the bound with mempty.

 trim   = splice mempty
 trim x = trimBefore _onset x . trimAfter _offset x

trimBefore :: Monoid a => Time -> Behavior a -> Behavior a

Replace everthing before the given time by mempty.

trimAfter :: Monoid a => Time -> Behavior a -> Behavior a

Replace everthing after the given time by mempty.

concatB :: Monoid a => Score (Behavior a) -> Behavior a

Concatenate a score of (possibly overlapping) segments.

See also concatSegment and continous.

Common behaviors

line :: Fractional a => Behavior a

A behavior that gives the current time, i.e. the identity function

Should really have the type Behavior Time, but is provided in a more general form for convenience.

unit :: Fractional a => Behavior a

A behavior that varies from 0 to 1 during the same time interval and is 0 before and 1 after that interval.

impulse :: Num a => Behavior a

A behavior that is 1 at time 0, and 0 at all other times.

turnOn :: Num a => Behavior a

A behavior that goes from 0 to 1 at time 0.

turnOff :: Num a => Behavior a

A behavior that goes from 1 to 0 at time 0.

sawtooth :: RealFrac a => Behavior a

A behavior that goes from 0 to 1 repeatedly with a period of 1.

sine :: Floating a => Behavior a

A behavior that

cosine :: Floating a => Behavior a

A behavior that

Segment type

data Segment a

A Segment is a value varying over some unknown time span.

To give a segment an explicit duration, use Stretched Segment.

To place a segment in a particular time span, use Note Segment.

The semantics are given by

 type Segment a = Duration -> a

Examples

segment :: Iso (Duration -> a) (Duration -> b) (Segment a) (Segment b)

View a segment as a time function and vice versa.

Combinators

focusing :: Lens' (Behavior a) (Segment a)

View part of a Behavior as a Segment.

 line & focusing `onSpan` (2 <-> 3) *~ 0

apSegments' :: Stretched (Segment a) -> Stretched (Segment a) -> Stretched (Segment a)

apSegments :: Voice (Segment a) -> Stretched (Segment a)

Append a voice of segments to a single stretched segment.

data Bound a

Bound restricts the start and stop time of a value, and prevents access to values outside the bounds.

Bound is especially useful to restrict the range of a Behavior. If we have a value with can only be reasonably defined for a particular time range, we can represent it as Bound Behavior. This is isomorphic to a Note Segment, and bounded whitnesses the isomorphism.

Bound is not Foldable or Traversable, as that would allow us to access values outside the bounds. However, we can still access values of a Bound Behavior in a safe manner using trim or splice.

The semantics are given by

 type Bound a = (Time, Time, a)

bounds :: Time -> Time -> a -> Bound a

Add bounds.

bounding :: Span -> a -> Bound a

Add bounds.

 (s,x)^.note = (bounding s . transform s) x

trim :: Monoid b => Bound (Behavior b) -> Behavior b

Extract a bounded behavior, replacing all values outside the bound with mempty.

 trim   = splice mempty
 trim x = trimBefore _onset x . trimAfter _offset x

splice :: Behavior a -> Bound (Behavior a) -> Behavior a

Inserts a bounded behavior on top of another behavior.

 trim = splice mempty

(Named after the analogous tape-editing technique.)

bounded' :: Iso' (Note (Segment a)) (Bound (Behavior a))

View a Note Segment as a Bound Behavior and vice versa.

This can be used to safely turn a behavior into a segment and vice versa. Often focusing is more convenient to use.

bounded :: Iso (Note (Segment a)) (Note (Segment b)) (Bound (Behavior a)) (Bound (Behavior b))

View a Note Segment as a Bound Behavior and vice versa.

This can be used to safely turn a behavior into a segment and vice versa. Often focusing is more convenient to use.