Music.Time.Stretched

Description

Stretched values

data Stretched a

A Stretched value has a known duration, but no position.

Placing a value inside Stretched makes it invariante under delay.

The semantics are given by

 type Stretched = (Duration, a)

Construction

stretched :: Iso (Duration, a) (Duration, b) (Stretched a) (Stretched b)

View a stretched value as a pair of the original value and a stretch factor.

stretchedValue :: (Transformable a, Transformable b) => Lens (Stretched a) (Stretched b) a b

View a stretched value as a pair of the original value and the transformation (and vice versa).