Music.Time.Time

Description

Documentation

Note that you should use .-. and .+^ to convert between time and duration. To refer to time zero (the beginning of the music), use origin.

Time type

type Time = Point Duration

This type represents absolute time in seconds since start. Note that time can be negative, representing events occuring before the start time.

Time forms an affine space with durations as the underlying vector space, that is, we can add a time to a duration to get a new time using .+^, take the difference of two times to get a duration using .-..

Duration type

data Duration

This type represents relative time in seconds.

Behavior class

class HasBehavior f where

Methods

(?) :: f a -> Time -> a

Identities

start :: Time

The global start time, which usually means the the beginning of the musical performance.

This is a synonym for origin.

stop :: Time

The global end time, defined as start .+^ unit.

unit :: Duration

The unit duration.

Combinators