Music.Time.Voice

Description

Voice type

data Voice a

A Voice is a sequential composition of values. Events may not overlap.

Construction

voice :: Getter [Stretched a] (Voice a)

Create a Voice from a list of Stretched values.

This is a Getter (rather than a function) for consistency:

[ (0 <-> 1, 10)^.stretched,
  (1 <-> 2, 20)^.stretched,
  (3 <-> 4, 30)^.stretched ]^.voice

view voice $ map (view stretched) [(0 <-> 1, 1)]

Se also stretcheds.

Extracting values

stretcheds :: Lens (Voice a) (Voice b) [Stretched a] [Stretched b]

View a Voice as a list of Stretched values.

view stretcheds                        :: Voice a -> [Stretched a]
set  stretcheds                        :: [Stretched a] -> Voice a -> Voice a
over stretcheds                        :: ([Stretched a] -> [Stretched b]) -> Voice a -> Voice b

preview  (stretcheds . each)           :: Voice a -> Maybe (Stretched a)
preview  (stretcheds . element 1)      :: Voice a -> Maybe (Stretched a)
preview  (stretcheds . elements odd)   :: Voice a -> Maybe (Stretched a)

set      (stretcheds . each)           :: Stretched a -> Voice a -> Voice a
set      (stretcheds . element 1)      :: Stretched a -> Voice a -> Voice a
set      (stretcheds . elements odd)   :: Stretched a -> Voice a -> Voice a

over     (stretcheds . each)           :: (Stretched a -> Stretched b) -> Voice a -> Voice b
over     (stretcheds . element 1)      :: (Stretched a -> Stretched a) -> Voice a -> Voice a
over     (stretcheds . elements odd)   :: (Stretched a -> Stretched a) -> Voice a -> Voice a

toListOf (stretcheds . each)                :: Voice a -> [Stretched a]
toListOf (stretcheds . elements odd)        :: Voice a -> [Stretched a]
toListOf (stretcheds . each . filtered
             (\x -> _duration x < 2))  :: Voice a -> [Stretched a]

This is not an Iso, as the note list representation does not contain meta-data. To construct a score from a note list, use score or flip (set stretcheds) empty.

eventsV :: Lens (Voice a) (Voice b) [(Duration, a)] [(Duration, b)]

singleStretched :: Prism' (Voice a) (Stretched a)

Deprecated: Use 'unsafeStretcheds . single'

Fusion

fuse :: Eq a => Voice a -> Voice a

Merge consecutive equal notes.

fuseBy :: (a -> a -> Bool) -> Voice a -> Voice a

Merge consecutive equal notes using the given function.

Fuse rests

fuseRests :: Voice (Maybe a) -> Voice (Maybe a)

Fuse all rests in the given voice. The resulting voice will have no consecutive rests.

coverRests :: Voice (Maybe a) -> Maybe (Voice a)

Remove all rests in the given voice by prolonging the previous note. Returns Nothing if and only if the given voice contains rests only.

Traversal

Separating rhythms and values

valuesV :: Lens (Voice a) (Voice b) [a] [b]

durationsV :: Lens' (Voice a) [Duration]

withValues :: ([a] -> [b]) -> Voice a -> Voice b

Transform the values, leaving durations intact.

withDurations :: ([Duration] -> [Duration]) -> Voice a -> Voice a

Transform the durations, leaving values intact.

rotateValues :: Int -> Voice a -> Voice a

Rotate values by the given number of steps, leaving durations intact.

rotateDurations :: Int -> Voice a -> Voice a

Rotate durations by the given number of steps, leaving values intact.

reverseValues :: Voice a -> Voice a

Reverse values, leaving durations intact.

reverseDurations :: Voice a -> Voice a

Reverse durations, leaving values intact.

Zips

unzipVoice :: Voice (a, b) -> (Voice a, Voice b)

Unzip the given voice. This is specialization of unzipR.

zipVoice :: Voice a -> Voice b -> Voice (a, b)

Join the given voices by multiplying durations and pairing values.

zipVoice3 :: Voice a -> Voice b -> Voice c -> Voice (a, (b, c))

Join the given voices by multiplying durations and pairing values.

zipVoice4 :: Voice a -> Voice b -> Voice c -> Voice d -> Voice (a, (b, (c, d)))

Join the given voices by multiplying durations and pairing values.

zipVoiceNoScale :: Voice a -> Voice b -> Voice (a, b)

Join the given voices by pairing values and selecting the first duration.

zipVoiceNoScale3 :: Voice a -> Voice b -> Voice c -> Voice (a, (b, c))

Join the given voices by pairing values and selecting the first duration.

zipVoiceNoScale4 :: Voice a -> Voice b -> Voice c -> Voice d -> Voice (a, (b, (c, d)))

Join the given voices by pairing values and selecting the first duration.

zipVoiceWith :: (a -> b -> c) -> Voice a -> Voice b -> Voice c

Join the given voices by multiplying durations and combining values using the given function.

zipVoiceWith' :: (Duration -> Duration -> Duration) -> (a -> b -> c) -> Voice a -> Voice b -> Voice c

Join the given voices by combining durations and values using the given function.

zipVoiceWithNoScale :: (a -> b -> c) -> Voice a -> Voice b -> Voice c

Join the given voices without combining durations.

Context

withContext :: Voice a -> Voice (Ctxt a)

voiceLens :: (s -> a) -> (b -> s -> t) -> Lens (Voice s) (Voice t) (Voice a) (Voice b)

voiceAsList :: Iso (Voice a) (Voice b) [a] [b]

listAsVoice :: Iso [a] [b] (Voice a) (Voice b)

Unsafe versions

unsafeStretcheds :: Iso (Voice a) (Voice b) [Stretched a] [Stretched b]

unsafeEventsV :: Iso (Voice a) (Voice b) [(Duration, a)] [(Duration, b)]