Music.Time.Meta

Description

Provides meta-information.

Each score supports an unlimited number of Reactive meta-values.

This is more or less based on Diagrams styles, which is in turn based on XMonad.

Attributes

type IsAttribute a = (Typeable a, Monoid a, Semigroup a)

data Attribute

An existential wrapper type to hold attributes.

Creating attributes

wrapAttr :: IsAttribute a => a -> Attribute

wrapTAttr :: (Transformable a, IsAttribute a) => a -> Attribute

unwrapAttr :: IsAttribute a => Attribute -> Maybe a

unwrapTAttr :: (Transformable a, IsAttribute a) => Attribute -> Maybe a

Meta-data

data Meta

Creating meta-data

toMeta :: forall a b. (IsAttribute b, Transformable b) => Maybe a -> b -> Meta

fromMeta :: forall a b. (IsAttribute b, Transformable b) => Maybe a -> Meta -> Maybe b

The HasMeta class

class HasMeta a where

Type class for things which have meta-information.

Methods

meta :: Lens' a Meta

Apply meta-information by combining it (on the left) with the existing meta-information.

applyMeta :: HasMeta a => Meta -> a -> a