Music.Time.Note

Description

Note type

data Note a

A value Note value, representing a value stretched by some Duration.

Construction

note :: Iso (Duration, a) (Duration, b) (Note a) (Note b)

View a note value as a pair of the original value and a stretch factor.

notee :: (Transformable a, Transformable b) => Lens (Note a) (Note b) a b

Access the note value. Taking a value out carries out the stretch (using the Transformable instance), while putting a value in carries out the reverse transformation.

>>> view notee $ (2,3::Duration)^.note
6

>>> set notee 6 $ (2,1::Duration)^.note
(2,3)^.note

durationNote :: Iso' Duration (Note ())

A note value as a duration carrying an associated value. Whitness by picking a trivial value.

>>> 2^.durationNote
(2,())^.note