Music.Time.Past

Documentation

newtype Past a

Past represents a value occuring before and at some point in time.

It may be seen as a note whose era is a left-open, right-inclusive time interval.

Constructors

Past

Fields

getPast :: (Min Time, a)

newtype Future a

Future represents a value occuring at and after some point in time.

It may be seen as a note whose era is a left-open, right-inclusive time interval.

Constructors

Future

Fields

getFuture :: (Max Time, a)

past :: Past a -> Time -> Maybe a

Query a past value. Semantic function.

future :: Future a -> Time -> Maybe a

Query a future value. Semantic function.

indexPast :: [Past a] -> Time -> Maybe a

firstTrue :: [Maybe a] -> Maybe a

pastSeg :: Past (Segment a) -> Behavior (Maybe a)

Project a segment (backwards) up to the given point.

futureSeg :: Future (Segment a) -> Behavior (Maybe a)

Project a segment starting from the given point.