Music.Time.Score

Description

Score type

data Score a

A Score is a sequential or parallel composition of values, and allows overlapping events

Query

Construction

score :: Getter [Event a] (Score a)

Create a score from a list of events.

events :: Lens (Score a) (Score b) [Event a] [Event b]

View a Score as a list of Event values.

eras :: Traversal' (Score a) Span

Print all eras of the given score.

>>> toListOf eras $ scat [c,d,e :: Score Integer]
[0 <-> 1,1 <-> 2,2 <-> 3]

triples :: Lens (Score a) (Score b) [(Time, Duration, a)] [(Time, Duration, b)]

View a score as a list of time-duration-value triplets.

Traversal

mapWithSpan :: (Span -> a -> b) -> Score a -> Score b

Map over the values in a score.

filterWithSpan :: (Span -> a -> Bool) -> Score a -> Score a

Filter the values in a score.

mapFilterWithSpan :: (Span -> a -> Maybe b) -> Score a -> Score b

Combination of mapTriples and filterTriples.

mapTriples :: (Time -> Duration -> a -> b) -> Score a -> Score b

Map over the values in a score.

filterTriples :: (Time -> Duration -> a -> Bool) -> Score a -> Score a

Filter the values in a score.

mapFilterTriples :: (Time -> Duration -> a -> Maybe b) -> Score a -> Score b

Efficient combination of mapTriples and filterTriples.

Simultaneous

hasOverlappingEvents :: Score a -> Bool

simultaneous :: (Transformable a, Semigroup a) => Score a -> Score a

Merge all simultaneous events using their Semigroup instance.

Normalize

normalizeScore :: Score a -> Score a

Normalize a score, assuring its events spans are all forward (as by isForwardSpan), and that its onset is at least zero. Consequently, the onset and offset of each event in the score is at least zero.

removeRests :: Score (Maybe a) -> Score a

Utility

printEras :: Score a -> IO ()

Print the span of each event, as given by eras.

Unsafe versions

unsafeEvents :: Iso (Score a) (Score b) [Event a] [Event b]

A score is a list of events up to meta-data. To preserve meta-data, use the more restricted score and events.

unsafeTriples :: Iso (Score a) (Score b) [(Time, Duration, a)] [(Time, Duration, b)]

A score is a list of (time-duration-value triples) up to meta-data. To preserve meta-data, use the more restricted triples.