Music.Time.Transform

Description

Documentation

module Music.Time.Types

The Transformable class

class Transformable a where

Class of values that can be transformed (i.e. scaled and moved) in time.

Law

 transform mempty = id
 transform (s <> t) = transform s . transform t
 transform (s <> negateV s) = id

Law

 onset (delay n a)       = n ^+. onset a
 offset (delay n a)      = n ^+. offset a
 duration (stretch n a)  = n * duration a
 duration (compress n a) = duration a / n

 delay n b ! t    = b ! (t .-^ n)
 undelay n b ! t  = b ! (t .+^ n)

Lemma

 duration a = duration (delay n a)

Methods

transform :: Span -> a -> a

itransform :: Transformable a => Span -> a -> a

Apply the inverse of the given transformation.

 itransform s = transform (negateV s)

transformed :: Transformable a => Span -> Iso' a a

View the given value in the context of the given transformation.

 over (transformed s) = (`whilst` s)

Apply under a transformation

whilst :: (Transformable a, Transformable b) => (a -> b) -> Span -> a -> b

Apply a function under transformation.

Designed to be used infix, as in

 stretch 2 `whilst` delaying 2

spanned :: (Transformable a, Transformable b) => Span -> Lens a b a b

Transforms a lens of to a Transformable type to act inside a transformation.

onSpan :: (Transformable a, Functor f) => LensLike' f a b -> Span -> LensLike' f a b

Transforms a lens of to a Transformable type to act inside a transformation.

Designed to be used infix, as in

 l `onSpan` (2 <-> 3)

Specific transformations

Transformations

delay :: Transformable a => Duration -> a -> a

Moves a value forward in time.

undelay :: Transformable a => Duration -> a -> a

Moves a value backward in time. Equnitvalent to stretch . negate.

stretch :: Transformable a => Duration -> a -> a

Stretches (augments) a value by the given factor.

compress :: Transformable a => Duration -> a -> a

Compresses (diminishes) a score. Equnitvalent to stretch . recip.

Transforming values

delaying :: Duration -> Span

A transformation that moves a value forward in time.

undelaying :: Duration -> Span

A transformation that moves a value backward in time.

stretching :: Duration -> Span

A transformation that stretches (augments) a value by the given factor.

compressing :: Duration -> Span

A transformation that compresses (diminishes) a value by the given factor.

Utility

delayTime :: Transformable a => Time -> a -> a

Delay relative to origin.

Provided for situations when we really want to use startAt, but the type does not have an instance for HasPosition and we can assume that the value is starting at time zero.