Music.Time.Types

Description

Duration

data Duration

Duration, corresponding to note values in standard notation. The standard names can be used: 1/2 for half note 1/4 for a quarter note and so on.

Duration is a one-dimensional VectorSpace, and is the associated vector space of time points. It is a also an AdditiveGroup (and hence also Monoid and Semigroup) under addition.

Duration is invariant under translation so delay has no effect on it.

The semantics are given by

 type Duration = R

toDuration :: Real a => a -> Duration

Convert a value to a duration.

fromDuration :: Fractional a => Duration -> a

Convert a value to a duration.

Time points

data Time

Time points, representing duration since some known reference time, typically the start of the music. Note that time can be negative, representing values occuring before the reference time.

Time forms an affine space with durations as the underlying vector space, that is, we can add a time to a duration to get a new time using .+^, take the difference of two times to get a duration using .-.. Time forms an AffineSpace with Duration as difference space.

The semantics are given by

 type Time = R

toTime :: Real a => a -> Time

Convert a value to a duration.

fromTime :: Fractional a => Time -> a

Convert a value to a duration.

Note that you should use .-. and .+^ to convert between time and duration. To refer to time zero (the beginning of the music), use origin.

Time spans

data Span

A Span represents an onset and offset in time (or equivalently: an onset and a duration, or a duration and an offset, or a duration and a middle point).

Pattern matching over span is possible (with ViewPatterns):

 foo (view range -> (t1, t2)) = ...
 foo (view delta -> (t, d)) = ...

Another way of looking at Span is that it represents a time transformation where onset is translation and duration is scaling. TODO see transform and whilst

The semantics are given by

 type Span = Time x Time

Creating spans

(<->) :: Time -> Time -> Span

t <-> u represents the span between t and u.

(>->) :: Time -> Duration -> Span

t >-> d represents the span between t and t .+^ d.

(<-<) :: Duration -> Time -> Span

d <-> t represents the span between t .-^ d and t.

Accessing spans

range :: Iso' Span (Time, Time)

View a span as pair of onset and offset.

delta :: Iso' Span (Time, Duration)

View a span as a pair of onset and duration.

showRange :: Span -> String

Show a span in range notation, i.e. t1 <-> t2.

showDelta :: Span -> String

Show a span in range notation, i.e. t >-> d.

Points in spans

isProper :: Span -> Bool

Whether this is a proper span, i.e. whether _onset x < stopTime x.

isBefore :: Span -> Span -> Bool

Whether the first given span occurs before the second span.

inside :: Time -> Span -> Bool

Whether the given point falls inside the given span (inclusively).

Designed to be used infix, for example

 0.5 `inside` (1 <-> 2)

encloses :: Span -> Span -> Bool

Whether the first given span encloses the second span.

overlaps :: Span -> Span -> Bool

Whether the given span overlaps.

Properties

delayOnly :: Prism' Span Time

A prism to the subset of Span that performs a delay but no stretch.

To access the delay component in any span, use view (delta . e_1)

stretchOnly :: Prism' Span Duration

A prism to the subset of Span that performs a stretch but no delay.

To access the stretch component in any span, use view (delta . _2)