Music.Time.Voice

Description

Voice type

data Voice a

A Voice is a sequential composition of non-overlapping note values.

Both Voice and Note have duration but no position. The difference is that Note sustains a single value throughout its duration, while a voice may contain multiple values. It is called voice because it is generalizes the notation of a voice in choral or multi-part instrumental music.

It may be useful to think about Voice and Note as vectors in time space (i.e. Duration), that also happens to carry around other values, such as pitches.

Construction

voice :: Getter [Note a] (Voice a)

Create a Voice from a list of Notes.

notes :: Lens (Voice a) (Voice b) [Note a] [Note b]

View a Voice as a list of Note values.

pairs :: Lens (Voice a) (Voice b) [(Duration, a)] [(Duration, b)]

View a score as a list of duration-value pairs. Analogous to triples.

durationsVoice :: Iso' [Duration] (Voice ())

Traversal

Separating rhythms and values

valuesV :: Lens (Voice a) (Voice b) [a] [b]

A lens to the values in a voice.

durationsV :: Lens' (Voice a) [Duration]

A lens to the durations in a voice.

Zips

unzipVoice :: Voice (a, b) -> (Voice a, Voice b)

Unzip the given voice.

zipVoiceScale :: Voice a -> Voice b -> Voice (a, b)

Join the given voices by multiplying durations and pairing values.

zipVoiceScale3 :: Voice a -> Voice b -> Voice c -> Voice (a, (b, c))

Join the given voices by multiplying durations and pairing values.

zipVoiceScale4 :: Voice a -> Voice b -> Voice c -> Voice d -> Voice (a, (b, (c, d)))

Join the given voices by multiplying durations and pairing values.

zipVoiceNoScale :: Voice a -> Voice b -> Voice (a, b)

Join the given voices by pairing values and selecting the first duration.

zipVoiceNoScale3 :: Voice a -> Voice b -> Voice c -> Voice (a, (b, c))

Join the given voices by pairing values and selecting the first duration.

zipVoiceNoScale4 :: Voice a -> Voice b -> Voice c -> Voice d -> Voice (a, (b, (c, d)))

Join the given voices by pairing values and selecting the first duration.

zipVoiceScaleWith :: (a -> b -> c) -> Voice a -> Voice b -> Voice c

Join the given voices by multiplying durations and combining values using the given function.

zipVoiceWithNoScale :: (a -> b -> c) -> Voice a -> Voice b -> Voice c

Join the given voices without combining durations.

zipVoiceWith' :: (Duration -> Duration -> Duration) -> (a -> b -> c) -> Voice a -> Voice b -> Voice c

Join the given voices by combining durations and values using the given function.

Fusion

fuse :: Eq a => Voice a -> Voice a

Merge consecutive equal notes.

fuseBy :: (a -> a -> Bool) -> Voice a -> Voice a

Merge consecutive notes deemed equal by the given predicate.

Fuse rests

fuseRests :: Voice (Maybe a) -> Voice (Maybe a)

Fuse all rests in the given voice. The resulting voice will have no consecutive rests.

coverRests :: Voice (Maybe a) -> Maybe (Voice a)

Remove all rests in the given voice by prolonging the previous note. Returns Nothing if and only if the given voice contains rests only.

Homophonic/Polyphonic texture

sameDurations :: Voice a -> Voice b -> Bool

Whether two notes have exactly the same duration pattern. Two empty voices are considered to have the same duration pattern. Voices with an non-equal number of notes differ by default.

mergeIfSameDuration :: Voice a -> Voice b -> Maybe (Voice (a, b))

Pair the values of two voices if and only if they have the same duration pattern (as per sameDurations).

mergeIfSameDurationWith :: (a -> b -> c) -> Voice a -> Voice b -> Maybe (Voice c)

Combine the values of two voices using the given function if and only if they have the same duration pattern (as per sameDurations).

homoToPolyphonic :: Voice [a] -> [Voice a]

Split a homophonic texture into a polyphonic one. The returned voice list will not have as many elements as the chord with the fewest number of notes.

Points in a voice

onsetsRelative :: Time -> Voice a -> [Time]

Returns the onsets of all notes in a voice given the onset of the first note.

offsetsRelative :: Time -> Voice a -> [Time]

Returns the offsets of all notes in a voice given the onset of the first note.

midpointsRelative :: Time -> Voice a -> [Time]

Returns the midpoints of all notes in a voice given the onset of the first note.

erasRelative :: Time -> Voice a -> [Span]

Returns the eras of all notes in a voice given the onset of the first note.

Context

withContext :: Voice a -> Voice (Ctxt a)

Decorate all notes in a voice with their context, i.e. previous and following value if present.

Unsafe versions

unsafeNotes :: Iso (Voice a) (Voice b) [Note a] [Note b]

A voice is a list of notes up to meta-data. To preserve meta-data, use the more restricted voice and notes.

unsafePairs :: Iso (Voice a) (Voice b) [(Duration, a)] [(Duration, b)]

A score is a list of (duration-value pairs) up to meta-data. To preserve meta-data, use the more restricted pairs.